+++ packages/nest/src/interceptors.ts
@@ -13,7 +13,7 @@ import {
   buildSecurityHeaders,
   type AuditService,
   type SecurityHeadersOptions,
-} from '@sabrina-shield/core';
+} from '@eksneks/core';
 import { METADATA, TOKENS } from './constants';
 import { toRequestContext, setResponseHeader } from './context';